backend/chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
sentence-transformers==2.3.1
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
